Refactoring the sprite editor
