+++ theme/monaco.less
@@ -1,6 +1,7 @@
 /* Import all components */
 @import 'themes/default/globals/site.variables';
 @import 'themes/pxt/globals/site.variables';
+@import 'spriteeditor.less';
 
 /* Reference import */
 @import (reference) "semantic.less";
