+++ theme/spriteeditor.less
@@ -166,7 +166,7 @@
 
 .monaco-field-editor-frame .closeIcon .close {
     position: absolute;
-    top: 1.0535rem;
+    top: .5rem;
     right: 1rem;
     color: rgba(0,0,0,.87);
     border-radius: 50%;
@@ -175,6 +175,7 @@
     height: 2.15rem;
     width: 2.15rem;
     padding: 0;
+    line-height: 34px;
 }
 
 .sprite-editor-glyph {
@@ -184,12 +185,23 @@
     // text-align: center;
     // border-right: 5px solid #4B7BEC;
     cursor: pointer;
-    background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3E%3Cstyle%3E.icon-canvas-transparent%7Bopacity:0;fill:%23f6f6f6%7D.icon-vs-out%7Bfill:%23f6f6f6%7D.icon-vs-bg%7Bfill:%23424242%7D.icon-vs-red%7Bfill:%23e51400%7D.icon-vs-yellow%7Bfill:%23fc0%7D.icon-vs-green%7Bfill:%23393%7D.icon-vs-blue%7Bfill:%231ba1e2%7D.icon-white%7Bfill:%23fff%7D.icon-vs-action-blue%7Bfill:%2300539c%7D%3C/style%3E%3Cpath class='icon-canvas-transparent' d='M16 16H0V0h16v16z' id='canvas'/%3E%3Cpath class='icon-vs-out' d='M16 8c0 4.411-3.589 8-8 8a2.803 2.803 0 0 1-2.8-2.8c0-.833.272-1.629.766-2.241a.596.596 0 0 0 .101-.359.667.667 0 0 0-.667-.666.58.58 0 0 0-.358.102c-.013.01-.029.017-.042.027v.271l-2.5 3.334L0 10.333V0h5v.589A7.945 7.945 0 0 1 8 0c4.411 0 8 3.589 8 8z' id='outline'/%3E%3Cpath class='icon-white' d='M8 2a5.958 5.958 0 0 0-3 .812v5.153A2.67 2.67 0 0 1 8.067 10.6a2.57 2.57 0 0 1-.544 1.614 1.593 1.593 0 0 0-.323.986.8.8 0 0 0 .8.8c3.309 0 6-2.691 6-6s-2.691-6-6-6z' id='iconFg'/%3E%3Cg id='iconBg'%3E%3Cpath class='icon-vs-blue' d='M7.125 3.625a.875.875 0 1 0 0 1.75.875.875 0 0 0 0-1.75z'/%3E%3Cpath class='icon-vs-green' d='M10.625 4.5a.875.875 0 1 0 0 1.75.875.875 0 0 0 0-1.75z'/%3E%3Cpath class='icon-vs-yellow' d='M11.5 8a.875.875 0 1 0 0 1.75.875.875 0 0 0 0-1.75z'/%3E%3Cpath class='icon-vs-red' d='M9.75 10.625a.875.875 0 1 0 0 1.75.875.875 0 0 0 0-1.75z'/%3E%3Cpath class='icon-vs-bg' d='M8 1c-1.075 0-2.089.25-3 .685v1.127A5.958 5.958 0 0 1 8 2c3.309 0 6 2.691 6 6s-2.691 6-6 6a.8.8 0 0 1-.8-.8c0-.372.118-.731.323-.986a2.573 2.573 0 0 0 .544-1.614A2.67 2.67 0 0 0 5 7.965v1.03c.13-.031.26-.062.4-.062A1.67 1.67 0 0 1 7.067 10.6c0 .378-.115.728-.323.987A2.572 2.572 0 0 0 6.2 13.2c0 .992.808 1.8 1.8 1.8 3.859 0 7-3.141 7-7 0-3.86-3.141-7-7-7z'/%3E%3C/g%3E%3Cg id='colorAction'%3E%3Cpath id='XMLID_16_' class='icon-vs-action-blue' d='M4 3H1V1h3v2zm0 1H1v5h3V4zm-1.5 8L4 10H1l1.5 2z'/%3E%3C/g%3E%3C/svg%3E");
-    background-position-x: 4px;
-    background-size: 85%;
+    background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3E%3Cstyle%3E.icon-canvas-transparent%7Bopacity:0;fill:%23f6f6f6%7D.icon-vs-out%7Bfill:%23f6f6f6%7D.icon-vs-bg%7Bfill:%23424242%7D.icon-vs-red%7Bfill:%23e51400%7D.icon-vs-yellow%7Bfill:%23ffcc00%7D.icon-vs-green%7Bfill:%23339933%7D.icon-vs-blue%7Bfill:%231ba1e2%7D.icon-vs-action-purple%7Bfill:%23652d90%7D.icon-white%7Bfill:%23ffffff%7D%3C/style%3E%3Cpath class='icon-canvas-transparent' d='M16 16H0V0h16v16z' id='canvas'/%3E%3Cpath class='icon-vs-out' d='M16 8c0 4.411-3.589 8-8 8a2.803 2.803 0 0 1-2.8-2.8c0-.833.272-1.629.766-2.241a.596.596 0 0 0 .101-.359.667.667 0 0 0-.667-.666.58.58 0 0 0-.358.102A3.584 3.584 0 0 1 2.8 10.8 2.803 2.803 0 0 1 0 8c0-4.411 3.589-8 8-8s8 3.589 8 8z' id='outline'/%3E%3Cpath class='icon-white' d='M5.4 7.933a2.67 2.67 0 0 1 2.667 2.666c0 .606-.193 1.179-.544 1.614a1.599 1.599 0 0 0-.323.987.8.8 0 0 0 .8.8c3.309 0 6-2.691 6-6s-2.691-6-6-6-6 2.691-6 6c0 .441.359.8.8.8.378 0 .729-.114.986-.322A2.568 2.568 0 0 1 5.4 7.933z' id='iconFg'/%3E%3Cg id='iconBg'%3E%3Cpath class='icon-vs-bg' d='M8 15c-.992 0-1.8-.808-1.8-1.8 0-.606.193-1.179.544-1.613.208-.259.323-.609.323-.987 0-.919-.748-1.666-1.667-1.666-.377 0-.728.115-.986.323A2.58 2.58 0 0 1 2.8 9.8C1.808 9.8 1 8.992 1 8c0-3.86 3.14-7 7-7 3.859 0 7 3.14 7 7 0 3.859-3.141 7-7 7zM5.4 7.933a2.67 2.67 0 0 1 2.667 2.666c0 .606-.193 1.179-.544 1.614a1.599 1.599 0 0 0-.323.987.8.8 0 0 0 .8.8c3.309 0 6-2.691 6-6s-2.691-6-6-6-6 2.691-6 6c0 .441.359.8.8.8.378 0 .729-.114.986-.322A2.568 2.568 0 0 1 5.4 7.933z'/%3E%3Cpath class='icon-vs-action-purple' d='M4.5 5.375a.875.875 0 1 0 0 1.75.875.875 0 0 0 0-1.75z'/%3E%3Cpath class='icon-vs-blue' d='M7.125 3.625a.875.875 0 1 0 0 1.75.875.875 0 0 0 0-1.75z'/%3E%3Cpath class='icon-vs-green' d='M10.625 4.5a.875.875 0 1 0 0 1.75.875.875 0 0 0 0-1.75z'/%3E%3Cpath class='icon-vs-yellow' d='M11.5 8a.875.875 0 1 0 0 1.75.875.875 0 0 0 0-1.75z'/%3E%3Cpath class='icon-vs-red' d='M9.75 10.625a.875.875 0 1 0 0 1.75.875.875 0 0 0 0-1.75z'/%3E%3C/g%3E%3C/svg%3E");
+    background-position: center;
+    background-size: 65%;
     background-repeat: no-repeat;
 }
 
-// .sprite-editor-glyph::before {
-//     content: "\f044"
-// }
+.sprite-focus-hover {
+    transition: all .15s ease-out;
+    border-radius: 50%;
+}
+
+.sprite-focus-hover:hover {
+    box-shadow: 0 0 0 3px rgba(0,0,0,.35);
+    -webkit-transform: scale(1.1);
+    transform: scale(1.1);
+}
+
+.sprite-editor-glyph.sprite-focus-hover:hover {
+    box-shadow: inset 0 0 0 2px rgba(0,0,0,.35);
+}
